# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Boilerplate",
    "summary": "A modular FastAPI starter with a plugin system",
    "description": "\n    # FastAPI Boilerplate\n\n    A modern FastAPI starter with:\n\n    * Vertical-slice modules and a clean infrastructure layer\n    * Session-based auth with OAuth providers\n    * Swappable cache, queue, and rate-limit backends\n    * SQLAdmin admin UI\n    ",
    "contact": {
      "name": "Benav Labs",
      "url": "https://github.com/benavlabs/FastAPI-boilerplate",
      "email": "contact@benav.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "User Login",
        "description": "Authenticates a user and creates a new session.\n\n            This endpoint accepts username/email and password credentials and verifies them.\n            On successful authentication:\n            - A new session is created\n            - A session ID is set as an HTTP-only cookie\n            - A CSRF token is generated for protection against CSRF attacks\n\n            The endpoint is protected by rate limiting to prevent brute force attacks.\n            After multiple failed attempts, further login attempts will be temporarily blocked.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful, session created",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or rate limit exceeded"
          },
          "429": {
            "description": "Too many login attempts, try again later"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "User Logout",
        "description": "Terminates the current user session.\n\n            This endpoint:\n            - Invalidates the active session in the storage backend\n            - Clears all session-related cookies from the client\n\n            After logout, the user will need to authenticate again to access\n            protected resources. Any existing session tokens will no longer be valid.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Logout successful, session terminated",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Logout Api V1 Auth Logout Post"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/auth/refresh-csrf": {
      "post": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "Refresh CSRF Token",
        "description": "Generates a new CSRF token for the current session.\n\n            This endpoint should be called to obtain a fresh CSRF token when:\n            - The current token is about to expire\n            - After a certain period of inactivity\n            - When increased security is needed for sensitive operations\n\n            The new token is returned in the response and also set as a cookie.",
        "operationId": "refresh_csrf_token_api_v1_auth_refresh_csrf_post",
        "responses": {
          "200": {
            "description": "New CSRF token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Refresh Csrf Token Api V1 Auth Refresh Csrf Post"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/auth/oauth/google": {
      "get": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "Initiate Google OAuth Login",
        "description": "Starts the OAuth 2.0 authentication flow with Google.\n\n            This endpoint generates the authorization URL that the user should be\n            redirected to in order to authenticate with Google. The flow includes:\n            - Creation of a state parameter for CSRF protection\n            - Generation of PKCE code challenge (for enhanced security)\n            - Setting appropriate OAuth scopes for profile access\n\n            After successful authentication with Google, the user will be redirected\n            back to this application's callback endpoint.\n\n            An optional redirect_uri can be specified to control where the user\n            is sent after the entire authentication process completes.",
        "operationId": "oauth_google_login_api_v1_auth_oauth_google_get",
        "parameters": [
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Redirect Uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Oauth Google Login Api V1 Auth Oauth Google Get"
                }
              }
            }
          },
          "500": {
            "description": "Failed to initiate Google login"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/callback/google": {
      "get": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "Google OAuth Callback Handler",
        "description": "Processes the authentication callback from Google OAuth.\n\n            This endpoint handles the authorization code returned by Google after\n            the user has successfully authenticated. The process includes:\n            - Validating the state parameter to prevent CSRF attacks\n            - Exchanging the authorization code for access/refresh tokens\n            - Fetching the user profile from Google\n            - Creating or updating the user account in the system\n            - Establishing a new session for the authenticated user\n\n            Two response formats are supported:\n            - redirect: Redirects to the frontend with success/error parameters (default)\n            - json: Returns user information and tokens as a JSON response\n\n            The json format is useful for mobile apps or single-page applications that\n            handle the OAuth flow programmatically.",
        "operationId": "oauth_google_callback_api_v1_auth_oauth_callback_google_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "response_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Response format, either 'redirect' or 'json'",
              "default": "redirect",
              "title": "Response Format"
            },
            "description": "Response format, either 'redirect' or 'json'"
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication successful (JSON response)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "302": {
            "description": "Authentication successful (redirect response)"
          },
          "400": {
            "description": "Invalid OAuth state or other parameter"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Server error during authentication"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/check-auth": {
      "get": {
        "tags": [
          "Authentication",
          "Authentication"
        ],
        "summary": "Check Auth",
        "description": "Check if the user is authenticated and return basic user information.\n\nThis is useful for clients to verify authentication status and can be used\nwith both cookie-based and API-based authentication.\n\nArgs:\n    session_data: The session data if the user is authenticated\n\nReturns:\n    Authentication status and user information if authenticated",
        "operationId": "check_auth_api_v1_auth_check_auth_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Auth Api V1 Auth Check Auth Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_api_v1_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_auth_login_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}